Almacena tweets en DB segun match con lucene

+++ src/main/java/cl/usach/traffictweet/Repositories/KeywordRepository.java
@@ -3,6 +3,6 @@ package cl.usach.traffictweet.Repositories;
 import cl.usach.traffictweet.Models.Keyword;
 import org.springframework.data.repository.CrudRepository;
 
-public interface KeywordRepository extends CrudRepository<Keyword,Long>{
+public interface KeywordRepository extends CrudRepository<Keyword,Integer>{
 
 }
