+++ src/main/java/cl/usach/traffictweet/Repositories/OccurrenceRepository.java
@@ -3,6 +3,6 @@ package cl.usach.traffictweet.Repositories;
 import cl.usach.traffictweet.Models.Occurrence;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OccurrenceRepository extends CrudRepository<Occurrence,Long>{
+public interface OccurrenceRepository extends CrudRepository<Occurrence,Integer>{
 
 }
